[suppressions] Refactor to improve testing
